requirements.txt
psycopg2
line-bot-sdk
opencv
python-dotenv
